--- usvm-ts/src/test/resources/samples/PtrCall.ts
@@ -1,9 +1,0 @@
-// @ts-nocheck
-// noinspection JSUnusedGlobalSymbols
-
-class PtrCall {
-    callLambda(): number {
-        const f = () => 42;
-        return f();
-    }
-}
